# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Mangaweb",
        "title": "Mangaweb3 API",
        "contact": {},
        "version": "3.0"
    },
    "paths": {
...
        "/browse/thumbnail": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the item",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
...
        "/tag/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tag.listRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tag.listResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/tag/set_favorite": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tag.setFavoriteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tag.setFavoriteResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/tag/thumbnail": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "tag",
                        "name": "tag",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
...
        "/view/download": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the file",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/view/fix_meta": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/view.fixMetaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/view.fixMetaResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
        "/view/get_image": {
            "get": {
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the item",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "width",
                        "name": "width",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "height",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "index",
                        "name": "i",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "body"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errors.Error": {
            "type": "object",
            "properties": {
                "cause": {},
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "message": {
                    "type": "string",
                    "example": "unknown error."
                }
            }
        },
...
        "tag.SortField": {
            "type": "string",
            "enum": [
                "name",
                "itemCount"
            ],
            "x-enum-varnames": [
                "SortFieldName",
                "SortFieldPageCount"
            ]
        },
        "tag.SortOrder": {
            "type": "string",
            "enum": [
                "ascending",
                "descending"
            ],
            "x-enum-varnames": [
                "SortOrderAscending",
                "SortOrderDescending"
            ]
        },
        "tag.Tag": {
            "type": "object",
            "properties": {
                "favorite": {
                    "type": "boolean"
                },
                "item_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "tag.listRequest": {
            "type": "object",
            "properties": {
                "favorite_only": {
                    "type": "boolean"
                },
                "item_per_page": {
                    "type": "integer",
                    "default": 30
                },
                "order": {
                    "$ref": "#/definitions/tag.SortOrder"
                },
                "page": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "sort": {
                    "$ref": "#/definitions/tag.SortField"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "tag.listResponse": {
            "type": "object",
            "properties": {
                "request": {
                    "$ref": "#/definitions/tag.listRequest"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tag.Tag"
                    }
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "tag.setFavoriteRequest": {
            "type": "object",
            "properties": {
                "favorite": {
                    "type": "boolean"
                },
                "tag": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "tag.setFavoriteResponse": {
            "type": "object",
            "properties": {
                "favorite": {
                    "type": "boolean"
                },
                "request": {
                    "$ref": "#/definitions/tag.setFavoriteRequest"
                }
            }
        },
        "view.fixMetaRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "view.fixMetaResponse": {
            "type": "object",
            "properties": {
                "request": {
                    "$ref": "#/definitions/view.fixMetaRequest"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
...
}